Fix to subfields() - was only processing one field.
Adds complexity to output of subfields().
Previous implementation also did not allow a way
to distinguish between, say, two 500 tags with 'a'
subfields and a single 500 tag with two 'a's.

Will need to add indicators to output as well, since
some display is indicator-value-dependent.  That change
will come in future patch.